Revise jinja2 loader
