+++ docs/gendocs.py
@@ -7,12 +7,14 @@ python gendocs.py
 """
 from __future__ import annotations
 
-import sys
 from pathlib import Path
 
 import mkdocs_gen_files
 from imas2xarray import Variable
+from jinja2 import Environment, FileSystemLoader
+from pydantic_yaml import to_yaml_str
 
+from duqtools.config import var_lookup
 from duqtools.config._schema_create import CreateConfigModel
 from duqtools.config._schema_root import ConfigModel
 from duqtools.ids._schema import ImasBaseModel
@@ -26,41 +28,91 @@ from duqtools.schema.data_location import DataLocation
 from duqtools.systems.jetto import JettoVar, JettoVariableModel
 from duqtools.systems.models import StatusConfigModel, SubmitConfigModel
 
-this_dir = Path(__file__).parent
-sys.path.append(str(this_dir))
+THIS_DIR = Path(__file__).parent
 
-from templates import get_template  # noqa
 
-objects = {
-    ARange,
-    ConfigModel,
-    CreateConfigModel,
-    DataLocation,
-    IDSOperationDim,
-    ImasBaseModel,
-    JettoVariableModel,
-    JettoVar,
-    LinSpace,
-    OperationDim,
-    StatusConfigModel,
-    SubmitConfigModel,
-    Variable,
-}
-schemas = {
-    f'schema_{obj.__name__}': obj.model_json_schema()  # type: ignore
-    for obj in objects
-}
-for page in (
-        'usage',
-        'systems_status',
-        'systems_submit',
-):
-    template = get_template(f'template_{page}.md')
+def get_template(path):
+    drc = path.parent
+    page = path.name
 
-    rendered = template.render(**schemas)
+    file_loader = FileSystemLoader(THIS_DIR / drc)
+    environment = Environment(loader=file_loader, autoescape=False)
+    environment.filters['to_yaml_str'] = to_yaml_str
 
-    filename = '/'.join(page.split('_')) + '.md'
+    return environment.get_template(page)
 
-    with mkdocs_gen_files.open(filename, 'w') as file:
+
+def gen_docs():
+    objects = {
+        ARange,
+        ConfigModel,
+        CreateConfigModel,
+        DataLocation,
+        IDSOperationDim,
+        ImasBaseModel,
+        JettoVariableModel,
+        JettoVar,
+        LinSpace,
+        OperationDim,
+        StatusConfigModel,
+        SubmitConfigModel,
+        Variable,
+    }
+    schemas = {
+        f'schema_{obj.__name__}': obj.model_json_schema()  # type: ignore
+        for obj in objects
+    }
+    for page in (
+            'usage.template.md',
+            'systems/status.template.md',
+            'systems/submit.template.md',
+    ):
+        path = Path(page)
+
+        template = get_template(path)
+
+        rendered = template.render(**schemas)
+
+        stem, *_ = path.name.partition('.')
+        new_path = path.with_name(stem).with_suffix('.md')
+
+        with mkdocs_gen_files.open(new_path, 'w') as file:
+            print(f'Writing {file.name}')
+            file.write(rendered)
+
+
+def gen_jetto_variables():
+    objects = {
+        JettoVariableModel,
+        JettoVar,
+    }
+    schemas = {
+        f'schema_{obj.__name__}': obj.model_json_schema()  # type: ignore
+        for obj in objects
+    }
+
+    path = Path('systems/variables.template.md')
+
+    template = get_template(path)
+
+    variable_groups = var_lookup.groupby_type()
+    variable_groups.pop('IDS-variable')
+
+    def sort_var_groups_in_dict(dct):
+        for name in dct:
+            dct[name] = sorted(dct[name], key=lambda var: var.name)
+
+    sort_var_groups_in_dict(variable_groups)
+
+    rendered = template.render(variable_groups=variable_groups, **schemas)
+
+    stem, *_ = path.name.partition('.')
+    new_path = path.with_name(stem).with_suffix('.md')
+
+    with mkdocs_gen_files.open(new_path, 'w') as file:
         print(f'Writing {file.name}')
         file.write(rendered)
+
+
+gen_docs()
+gen_jetto_variables()
