+++ docs/systems/variables.template.md
